if we are calling get all claims api with params, in this case i am not saving the api data in local db.

See merge request cilalabs/claimguru/claimguru-ui!648

+++ src/store/claims/mutations.js
@@ -2,14 +2,16 @@ import localDB, { getCollection } from '@services/dexie';
 
 export async function setClaims(state, claimsData = []) {
   const claimsCollection = await getCollection('claims');
-  const claims = claimsData.map(claim => ({
+  const claims = claimsData.data.map(claim => ({
     ...claim
   }));
   state.claims = claims;
-  if ((await claimsCollection.count()) > 0) {
+  if ((await claimsCollection.count()) > 0 && !claimsData.params) {
     await claimsCollection.delete([]);
   }
-  await localDB.claims.bulkAdd(claims);
+  if (!claimsData.params) {
+    await localDB.claims.bulkAdd(claims);
+  }
 }
 
 export async function setOfflineClaims(state) {
